Moved FirebaseAppScheduler out of the interface file
